rename daemon and remove websocket
