+++ pkg/service/api/readers.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/go-chi/render"
 	"github.com/rs/zerolog/log"
-	"github.com/wizzomafizzo/tapto/pkg/daemon/state"
+	"github.com/wizzomafizzo/tapto/pkg/service/state"
 )
 
 type ReaderWriteRequest struct {
